Qr App/requirements.txt
# Core dependencies
tkinter
pillow
pyperclip

# QR code handling
qrcode
pyzbar
opencv-python

# Other utilities
numpy